make old class work, put common code into base class
