+++ src/main/scala/com/amazon/deequ/anomalydetection/RateOfChangeStrategy.scala
@@ -1,10 +1,29 @@
-package com.amazon.deequ.anomalydetection
+/**
+ * Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"). You may not
+ * use this file except in compliance with the License. A copy of the License
+ * is located at
+ *
+ *     http://aws.amazon.com/apache2.0/
+ *
+ * or in the "license" file accompanying this file. This file is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ *
+ */
 
+
+package com.amazon.deequ.anomalydetection
 
 /**
   * Provided for backwards compatibility.
   * the old [[RateOfChangeStrategy]] actually detects absolute changes so it has been migrated to [[AbsoluteChangeStrategy]]
   * use [[RelativeRateOfChangeStrategy]] if you want to detect relative changes to the previous values
   */
-class RateOfChangeStrategy extends AbsoluteChangeStrategy {
-}
+@deprecated
+case class RateOfChangeStrategy(
+  maxRateDecrease: Option[Double] = None,
+  maxRateIncrease: Option[Double] = None,
+  order: Int = 1) extends BaseChangeStrategy
